# packages/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Yellow House MVP API",
    "description": "Schedule coordination app API for finding overlapping availability across groups",
    "version": "1.0.0",
    "contact": {
      "name": "Mountaincoders",
      "url": "https://mountaincoders.ch"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.yellowhouse.app",
      "description": "Production server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check endpoint",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "email": { "type": "string" },
                        "name": { "type": "string" },
                        "created_at": { "type": "string", "format": "date-time" }
                      }
                    },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "User already exists or invalid input" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "email": { "type": "string" },
                        "name": { "type": "string" }
                      }
                    },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout user (requires authentication)",
        "tags": ["Authentication"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Logout successful" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/request-reset": {
      "post": {
        "summary": "Request password reset",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reset email sent (or user not found for security)" }
        }
      }
    },
    "/auth/confirm-reset": {
      "post": {
        "summary": "Confirm password reset",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "newPassword"],
                "properties": {
                  "token": { "type": "string" },
                  "newPassword": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset successful" },
          "400": { "description": "Invalid or expired token" }
        }
      }
    },
    "/groups": {
      "post": {
        "summary": "Create a new group",
        "tags": ["Groups"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Group created" },
          "401": { "description": "Unauthorized" }
        }
      },
      "get": {
        "summary": "Get user's groups",
        "tags": ["Groups"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of user's groups" }
        }
      }
    },
    "/groups/{groupId}": {
      "get": {
        "summary": "Get group details",
        "tags": ["Groups"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Group details" },
          "404": { "description": "Group not found" }
        }
      }
    },
    "/groups/{groupId}/members": {
      "get": {
        "summary": "Get group members",
        "tags": ["Groups"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "List of group members" }
        }
      }
    },
    "/groups/{groupId}/join": {
      "post": {
        "summary": "Join a group",
        "tags": ["Groups"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Successfully joined group" }
        }
      }
    },
    "/groups/{groupId}/availability": {
      "post": {
        "summary": "Mark availability for a time slot",
        "tags": ["Availability"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["time_slot", "marked"],
                "properties": {
                  "time_slot": { "type": "string", "example": "2024-03-02T09:00:00" },
                  "marked": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Availability marked" }
        }
      },
      "get": {
        "summary": "Get group's availability slots",
        "tags": ["Availability"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "List of availability slots" }
        }
      }
    },
    "/groups/{groupId}/overlaps": {
      "get": {
        "summary": "Get availability overlaps for group",
        "tags": ["Availability"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Filter overlaps from this date (YYYY-MM-DD)"
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date" },
            "description": "Filter overlaps until this date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": { "description": "List of overlaps with percentages" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}